requirements.txt
Django==5.0.6
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-decouple==3.8
reportlab==4.0.7
PyPDF2==3.0.1 